chore: organize components by feature
